First tagging solution added

+++ items/picture.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
-from auth import pygal_user
 from items import base_item_props
 from items import itemlist
 from pylibs import osm
 from app import piclink
+from app import prefix_add_tag
+from app import prefix_delete
 from app import prefix_info
 from app import prefix_thumbnail
 from app import prefix_webnail
@@ -20,13 +21,102 @@ import pygal_config as config
 import time
 
 
-class picture(base_item_props):
+class tags(dict):
+    def __init__(self):
+        self._id = 0
+        try:
+            dict.__init__(self, json.loads(open(self.tag_path(), 'r').read()))
+            for ident in self:
+                if int(ident) > self._id:
+                    self._id = int(ident)
+        except:
+            dict.__init__(self)
+
+    def tag_id_exists(self, tag_id):
+        return tag_id in self
+
+    def get_tag_id_list(self):
+        return self.keys()
+
+    def get_tag_wn_x(self, tag_id):
+        try:
+            return int(self[tag_id]['x'] * self.webnail_x())
+        except:
+            return ''
+
+    def get_tag_wn_y(self, tag_id):
+        try:
+            return int(self[tag_id]['y'] * self.webnail_y())
+        except:
+            return ''
+
+    def get_tag_wn_w(self, tag_id):
+        try:
+            return int(self[tag_id]['w'] * self.webnail_x())
+        except:
+            return ''
+
+    def get_tag_wn_h(self, tag_id):
+        try:
+            return int(self[tag_id]['h'] * self.webnail_y())
+        except:
+            return ''
+
+    def get_tag_wn_x2(self, tag_id):
+        try:
+            return self.get_tag_wn_x(tag_id) + self.get_tag_wn_w(tag_id)
+        except:
+            return ''
+
+    def get_tag_wn_y2(self, tag_id):
+        try:
+            return self.get_tag_wn_y(tag_id) + self.get_tag_wn_h(tag_id)
+        except:
+            return ''
+
+    def get_tag_text(self, tag_id):
+        try:
+            return self[tag_id]['tag']
+        except:
+            return ''
+
+    def add_tag_wn_xywh(self, x, y, w, h, tag, ident=None):
+        tag_dict = dict()
+        tag_dict['x'] = float(x) / self.webnail_x()
+        tag_dict['y'] = float(y) / self.webnail_y()
+        tag_dict['w'] = float(w) / self.webnail_x()
+        tag_dict['h'] = float(h) / self.webnail_y()
+        tag_dict['tag'] = tag
+        if ident is None:
+            self._id += 1
+            self[str(self._id)] = tag_dict
+        else:
+            self[str(ident)] = tag_dict
+        self._save_tags()
+
+    def add_tag_wn_x1y1x2y2(self, x1, y1, x2, y2, tag_text, tag_id=None):
+        self.add_tag_wn_xywh(min(x1, x2), min(y1, y2), abs(x2-x1), abs(y2-y1), tag_text, tag_id)
+
+    def delete_tag(self, tag_id):
+        if self.tag_id_exists(tag_id):
+            del self[tag_id]
+            self._save_tags()
+
+    def _save_tags(self):
+        if self.tag_path() is not None:
+            with open(self.tag_path(), 'w') as fh:
+                fh.write(json.dumps(self, indent=3, sort_keys=True))
+
+
+
+class picture(base_item_props, tags):
     mime_types = {'jpg': 'image/jpeg', 'jpeg': 'image/jpeg', 'jpe': 'image/jpeg', 'png': 'image/png', 'tif': 'image/tiff', 'tiff': 'image/tiff', 'gif': 'image/gif'}
     required_prop_keys = ['raw_x', 'raw_y', 'time', 'orientation', 'manufactor', 'model']
     prop_vers = 0.1
 
     def __init__(self, rel_path, request_args={}, prefix='', parent=None, **kwargs):
         base_item_props.__init__(self, rel_path, request_args=request_args, prefix=prefix, parent=parent)
+        tags.__init__(self)
         logger.debug('Initialising %s', self.name(True))
         self._info = picture_info_cached(self.raw_path(), self.prop_item_path(), logger=logger)
         self._citem_info = None
@@ -53,6 +143,8 @@ class picture(base_item_props):
         # rv.append(piclink(self.edit_url(), 'Edit', config.url_prefix + '/static/pygal_theme/img/edit.png'))
         if self._prefix != prefix_info:
             rv.append(piclink(self.info_url(), 'Info', config.url_prefix + '/static/common/img/info.png'))
+        if self._prefix != prefix_add_tag:
+            rv.append(piclink(self.add_tag_url(), 'Add Tag', config.url_prefix + '/static/common/img/edit.png'))
         if self.user_may_download():
             rv.append(piclink(self.download_url(), 'Download', config.url_prefix + '/static/common/img/download.png'))
         if self.gps() is not None:
@@ -62,7 +154,7 @@ class picture(base_item_props):
             rv.append(piclink(self.url(), 'Stop Slideshow', config.url_prefix + '/static/common/img/stop_slideshow.png'))
         else:
             rv.append(piclink(self.slideshow_url(), 'Start Slideshow', config.url_prefix + '/static/common/img/start_slideshow.png'))
-        if self.user_may_delete():
+        if self.user_may_delete() and self._prefix != prefix_delete:
             rv.append(piclink(self.delete_url(), 'Delete', config.url_prefix + '/static/common/img/delete.png'))
         return rv
